requirements.txt
fastapi
uvicorn
pydantic
google-genai
google-auth
google-cloud-aiplatform
python-dotenv
